requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.6.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
control==0.9.2
cryptography==37.0.4
cssselect==1.1.0
cycler==0.11.0
distlib==0.3.4
dj-database-url==1.2.0
dj-static==0.0.6
Django==4.1.6
django-bootstrap4==22.3
django-heroku==0.3.1
django-stdimage==6.0.1
emoji==2.0.0
et-xmlfile==1.1.0
fake-useragent==0.1.11
filelock==3.7.1
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
fonttools==4.34.4
frozenlist==1.3.0
future==0.18.2
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.47.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mpmath==1.2.1
multidict==6.0.2
neat-python==0.92
neo==0.6.1
nltk==3.7
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parse==1.19.0
pefile==2022.5.30
Pillow==9.2.0
platformdirs==2.5.2
protobuf==3.19.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDAQmx==1.4.6
pyee==8.2.2
pygame==2.1.2
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyqtTools==0.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyquery==1.4.3
PySocks==1.7.1
pyTelegramBotAPI==4.7.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
quantities==0.13.0
regex==2022.7.25
requests==2.7.0
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
seaborn==0.11.2
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.3
static3==0.7.0
sympy==1.11.1
telegram-send==0.34
Telethon==1.24.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tflearn==0.5.0
tornado==6.2
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.10
virtualenv==20.15.1
virtualenvwrapper-win==1.2.7
visualize==0.5.1
w3lib==1.22.0
websockets==10.3
Werkzeug==2.1.2
whitenoise==6.3.0
wrapt==1.14.1
wsproto==1.1.0
xvfbwrapper==0.2.9
yarl==1.7.2
zipp==3.8.1
